requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
fastjsonschema==2.21.2
filelock==3.24.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
more-itertools==10.8.0
msgpack==1.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
platformdirs==4.9.2
pyaml==26.2.1
pyarrow==23.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zstandard==0.25.0
